Rename "strategy" and add "auto detect source type" system
